requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==22.1.0 #need
certifi==2020.12.5
chardet==5.0.0
packaging==19.2
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
requests==2.28.1
selenium==4.5.0
six==1.13.0
urllib3==1.25.9
webdriver-manager==3.8.4
zipp==0.6.0
beautifulsoup4==4.11.1
google-api-python-client==2.65.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.0
pathlib2==2.3.7
pytest-base-url==2.0.0

